Refactoring
Move position form into position folder
