--- src/addPosition/changeValue.js
@@ -1,5 +1,0 @@
-const changeValue = ({ dispatch, value, key}) => (
-  dispatch({ type: 'ADD_POSITION_VALUE_CHANGED', value, key })
-);
-
-export default changeValue;
