--- src/addPosition/closeModal.js
@@ -1,2 +1,0 @@
-const closeModal = ({dispatch}) => dispatch({ type: 'CLOSE_ADD_MODAL' });
-export default closeModal;
